make len a getter
Convert the `len` member to `_len`.
make `len` a getter function.

Why?
I'm looking to change the way indices are handled internally.
this should make it unnecessary to have a `len` field.
Instead, the length will be calculated from the indices.
So I'm making it a function in preperation for the change.